[KYUUBI #4876] fix Flaky test BackendServiceMetricSuite
### _Why are the changes needed?_

### _How was this patch tested?_
- [ ] Add some test cases that check the changes thoroughly including negative and positive cases if possible

- [ ] Add screenshots for manual tests if appropriate

- [ ] [Run test](https://kyuubi.readthedocs.io/en/master/develop_tools/testing.html#running-tests) locally before make a pull request

Closes #4926 from lightning-L/branch-1.7_cp.

Closes #4876

0d8f5e2e3 [Tianlin Liao] [KYUUBI #4876] fix Flaky test BackendServiceMetricSuite

Authored-by: Tianlin Liao <tiliao@ebay.com>

+++ kyuubi-server/src/test/scala/org/apache/kyuubi/server/BackendServiceMetricSuite.scala
@@ -69,7 +69,7 @@ class BackendServiceMetricSuite extends WithKyuubiServer with HiveJDBCTestHelper
       assert(
         timer2.get(MetricsConstants.BS_OPEN_SESSION).get("count").asInt() == 1)
       assert(
-        timer2.get(MetricsConstants.BS_OPEN_SESSION).get("min").asInt() > 0)
+        timer2.get(MetricsConstants.BS_OPEN_SESSION).get("min").asDouble() > 0)
       val execStatementNode2 = timer2.get(MetricsConstants.BS_EXECUTE_STATEMENT)
       assert(execStatementNode2.get("count").asInt() == 3)
       assert(
